fix: fixed the `drawFilledRect` function
That function was bugged as hell, mainly because I was trying to draw in
chunks. When I tried to use it for backspace I had all sorts of weird
visual glitches as I'd have rects drawing in incorrect places and such.
Anyways, I simplified it and got rid of all chunkingg and now it works
as expected. And yes I added backspace support for deleting previous
characters.