chore: video driver further cleaning
just making sure to avoid using unncessary types and avoiding `g.c.` or
whatever prefixes whenever possible. it's surprising how overcomplicated
i made my code, like a bunch of variables have overly complicated types
that can be replaced with something far simpler with the help of the
`undefined` keyword whenever i need an uninitialized variable. for
example, it serves as a placeholder value for the `fb` variable which
holds the variable in global scope so all functions in the video driver
can use it. however, it's not actually initialized with the correct
framebuffer until the init function is called of course, so i use
`undefined` as a placeholder until i can set it to point at the correct
framebuffer.