chore: updated tiny printf
i decided to use a fork of tiny-printf which is about 300+ commits ahead
of the original repo that i was using! the original repo was written by
mpaland (marco paland) while this fork is maintained by eyalroz (eyal
rozenberg). this fork has a bunch of bug fixes and extra features and
such so might as well use it no? again, this is like 300+ commits ahead
of the original repo which i was using. the original repo is relatively
unmaintained anyways.